.github/wordcloud/requirements.txt
wordcloud==1.9.2
matplotlib==3.7.2
numpy==1.24.3
Pillow>=10.3.0  # Security fix for CVE-2023-50447
